Revert "Change: (listen-queue--update-buffer) Add REVERTP argument"
This reverts commit 66a220243a135aa2e82de14631b57f590011b4e0.

It's necessary to revert the queue to draw the arrow next to the
current track, which is used to find where to make the overlay.

(Attempts at drawing the arrow without reverting the vtable were
unsatisfying due to vtable's not getting to take the arrow's width
into account while drawing the table if it's not present when
inserting the row.  And attempts to just replace the two rows of the
old/current track failed due to a weird problem, maybe a bug in
vtable, not sure.)