WIP: fade mode
This isn't working because apparently VLC does not accurately return
its position near the end of a track (as it gets within 15 seconds or
so, it just skips back to the beginning of the track, so the time
remaining never shows less than 10-15 seconds, making it difficult to
start fading that close to the end of the track).  Another approach
will be needed, I guess; or perhaps another player backend would work
better for this feature.