WIP: Replace queue tracks
Not sure if I really like this solution.  I think it might be better
to have a special queue, maybe called "*now*", which "play" commands
use by default, and which has its tracks replaced by default.  Then
to build a playlist of tracks, users could choose to add tracks to a
specific queue, and then they could play that queue when ready.