WIP
This mostly works, but there are still things to fix.  For example:

- When the player is paused, the duration in the lighter still counts
down.  Will have to store the position in a slot and update that, and
calculate remaining from that; and stop updating the position when the
player is paused; and when the player is unpaused, set the started-at
and started-from accordingly.  Ugh.  This is why I tried to avoid
caching this information and just fetch it from the player every time.